+++ build.gradle
@@ -4,6 +4,7 @@ plugins {
 	id 'org.springframework.boot' version '3.4.1'
 	id 'io.spring.dependency-management' version '1.1.7'
 	id 'org.jetbrains.kotlin.plugin.jpa' version '1.9.25'
+	id 'org.jlleitschuh.gradle.ktlint' version '12.1.2'
 }
 
 group = 'com.example'
