--- src/main/kotlin/com/example/power_clean/PowerCleanController.kt
@@ -1,13 +1,0 @@
-package com.example.power_clean
-
-import org.springframework.web.bind.annotation.RestController
-import org.springframework.web.bind.annotation.GetMapping
-
-
-@RestController
-class PowerCleanController {
-    @GetMapping("/health-check")
-    fun healthCheck(): String {
-        return "I'm healthy!"
-    }
-}
