--- src/main/kotlin/com/example/power_clean/domain/repository/PostRepository.kt
@@ -1,3 +1,0 @@
-package com.example.power_clean.domain.repository
-
-public interface PostRepository {}
