--- src/main/kotlin/com/example/power_clean/presentation/rest/AccountController.kt
@@ -1,10 +1,0 @@
-package com.example.power_clean.presentation.rest
-
-import org.springframework.web.bind.annotation.RestController
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.security.core.annotation.AuthenticationPrincipal
-import org.springframework.security.oauth2.core.user.OAuth2User
-
-@RestController("account")
-class AccountController {
-}
