+++ src/main/kotlin/com/example/powerclean/PowerCleanApplication.kt
@@ -1,0 +1,11 @@
+package com.example.powerclean
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication()
+class PowerCleanApplication
+
+fun main(args: Array<String>) {
+    runApplication<PowerCleanApplication>(*args)
+}
