+++ src/main/kotlin/com/example/powerclean/domain/model/Book.kt
@@ -1,0 +1,28 @@
+package com.example.powerclean.domain.model
+
+import jakarta.persistence.Column
+import jakarta.persistence.ConstraintMode
+import jakarta.persistence.Embedded
+import jakarta.persistence.Entity
+import jakarta.persistence.ForeignKey
+import jakarta.persistence.JoinColumn
+import jakarta.persistence.OneToOne
+import jakarta.persistence.Table
+
+@Entity
+@Table(name = "book")
+class Book(
+    @Column(name = "title", nullable = false)
+    var title: String,
+    @Column(name = "content", nullable = false)
+    var content: String,
+    @Column(name = "link", nullable = false)
+    var link: String,
+    @Embedded
+    var authorInfo: AuthorInfo,
+    @Column(name = "post_id", nullable = false)
+    var postId: Long,
+    @OneToOne()
+    @JoinColumn(name = "post_id", foreignKey = ForeignKey(ConstraintMode.NO_CONSTRAINT))
+    var post: Post,
+) : BaseEntity()
