+++ src/main/kotlin/com/example/powerclean/domain/repository/AccountRepository.kt
@@ -1,0 +1,9 @@
+package com.example.powerclean.domain.repository
+
+import com.example.powerclean.domain.model.Account
+
+public interface AccountRepository {
+    fun save(account: Account): Account
+
+    fun findByName(name: String): Account?
+}
