+++ src/main/kotlin/com/example/powerclean/domain/repository/BookRepository.kt
@@ -1,0 +1,3 @@
+package com.example.powerclean.domain.repository
+
+public interface BookRepository
