+++ src/main/kotlin/com/example/powerclean/domain/repository/OauthProfileRepository.kt
@@ -1,0 +1,7 @@
+package com.example.powerclean.domain.repository
+
+import com.example.powerclean.domain.model.OauthProfile
+
+public interface OauthProfileRepository {
+    fun save(oauthProfile: OauthProfile): OauthProfile
+}
