+++ src/main/kotlin/com/example/powerclean/domain/repository/PostRepository.kt
@@ -1,0 +1,3 @@
+package com.example.powerclean.domain.repository
+
+public interface PostRepository
