+++ src/main/kotlin/com/example/powerclean/domain/repository/ReviewRepository.kt
@@ -1,0 +1,3 @@
+package com.example.powerclean.domain.repository
+
+public interface ReviewRepository
