+++ src/main/kotlin/com/example/powerclean/presentation/rest/AccountController.kt
@@ -1,0 +1,6 @@
+package com.example.powerclean.presentation.rest
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController("account")
+class AccountController
