--- src/test/kotlin/com/example/power_clean/PowerCleanApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.power_clean
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class PowerCleanApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
