+++ src/test/kotlin/com/example/powerclean/PowerCleanApplicationTests.kt
@@ -1,0 +1,11 @@
+package com.example.powerclean
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class PowerCleanApplicationTests {
+    @Test
+    fun contextLoads() {
+    }
+}
